% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_check}
\alias{dir_check}
\title{Check directory}
\usage{
dir_check(path, min_depth = 2)
}
\arguments{
\item{path}{character value specifying a directory.}

\item{min_depth}{numeric value indicating the minimum depth.}
}
\description{
Checks if a directory exists, creating folder(s) if necessary.
}
\details{
If the first min_depth folders in the directory do not exist, there are likely typos in \code{path} and an error will be thrown.
}
\author{
Jireh Huang (\email{jirehhuang@ucla.edu})
}
