% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{build_data_grid}
\alias{build_data_grid}
\title{Build data grid.}
\usage{
build_data_grid(
  network = "survey",
  var_lb = 0.5,
  var_ub = 1,
  coef_lb = 0.5,
  coef_ub = 1,
  normalize = TRUE,
  copies = 1,
  debug = FALSE
)
}
\arguments{
\item{network}{character value indicating network name.}

\item{debug}{logical value that activates printing debugging output.}
}
\value{
data.frame with grid of network settings.
}
\description{
Builds and formats a data grid of simulations.
}
\examples{
## Default data_grid
build_data_grid()
}
\author{
Jireh Huang (\email{jirehhuang@ucla.edu})
}
